Update unifi library to patched version to fix auth issue
Updates the node-unifi version to a patched version of 1.4.3 (numbered 1.4.4 in my forked github repo) which backports a bugfix impacting the proper execution of all commands.

In a recent verison of unifi network release, the CSRF token now changes with each command. This patch tracks the CSRF token properly so commands continue to execute correctly.

Note: using a forked patch b/c the library’s patch is on the 2.x version of the module which has other major changes which would require a significant refactor.  We’ll need to update the dependency to the 2.x version long term but wanted to get this module working again in the mean time.